+++ R/curvep.R
@@ -78,6 +78,7 @@ get_monotonics	<- function (vals, vdif = 0, bads = NULL)
 #' @param StrictImp It prevents extrapolating over concentration-range boundaries; used for POD, ECxx etc (default = TRUE).
 #' @param DUMV A dummy value, default = -999.
 #' @param TLOG A scaling factor for calculating the wAUC, default = -24.
+#' @param ... allow other parameters to pass
 #'
 #' @return A list with corrected concentration-response measurements and several calculated curve metrics.
 #' \itemize{
@@ -95,6 +96,7 @@ get_monotonics	<- function (vals, vdif = 0, bads = NULL)
 #'   \item nCorrected: number of points corrected (basically, sum of flags in corr)
 #'   \item Comments: warning and notes about the dose-response curve
 #'   \item Settings: input parameters for this run
+#'
 #' }
 #' @export
 #' @references{
@@ -113,7 +115,7 @@ get_monotonics	<- function (vals, vdif = 0, bads = NULL)
 #[[Rccp::export]]
 curvep <- function(Conc, Resp, Mask = NULL,
                     TRSH = 15, RNGE = -100, MXDV = 5, CARR = 0, BSFT = 3, USHP = 4,
-                    TrustHi = FALSE, StrictImp = TRUE, DUMV = -999, TLOG = -24)
+                    TrustHi = FALSE, StrictImp = TRUE, DUMV = -999, TLOG = -24, ...)
 # Conc - array of concentrations, e.g., in Molar units, can be log-transformed, in which case internal log-transformation is skipped
 # Resp - array of responses at corresponding concentrations, e.g., raw measurements or normalized to controls
 # !!!NB: Conc & Resp arrays should be in order from lo to hi concentrations