+++ R/rcurvep_combirun.R
@@ -27,16 +27,16 @@
 #'
 #' # create the zfishdev_act dataset
 #'\donttest{
-#'  set.seed(300)
+#'
 #'  data(zfishdev_all)
 #'  zfishdev_act <- combi_run_rcurvep(
 #'    zfishdev_all, n_samples = 100, keep_sets = c("act_set"),TRSH = seq(5, 95, by = 5),
-#'    RNGE = 1000000, CARR = 20
+#'    RNGE = 1000000, CARR = 20, seed = 300
 #'  )
 #'}
 #'
 combi_run_rcurvep <- function(d, n_samples = NULL, vdata = NULL, mask = 0,
-                              keep_sets = c("act_set", "resp_set", "fp_set"), ...) {
+                              keep_sets = c("act_set", "resp_set", "fp_set"),  ...) {
 
   paras <- list(...)
 
@@ -52,6 +52,7 @@ combi_run_rcurvep <- function(d, n_samples = NULL, vdata = NULL, mask = 0,
   keep_sets <- .check_keep_sets(keep_sets, c("act_set", "resp_set", "fp_set"), must_set = "act_set")
 
   # create inputs
+  if (!is.na(new_config$seed)) set.seed(new_config$seed)
   d1 <- create_dataset(d, n_samples = n_samples, vdata = vdata)
   para_in <- create_para_input(paras, n_samples = n_samples, d = d1)
 
@@ -96,7 +97,8 @@ merge_rcurvep_output <- function(d, keep_sets) {
 
   result <- purrr::map(
     keep_sets, function(x, tbl_names, obj)
-      suppressWarnings(obj[c("endpoint", "chemical", tbl_names[[x]])] %>% tidyr::unnest()),
+      obj[c("endpoint", "chemical", tbl_names[[x]])] %>%
+      tidyr::unnest(cols = tbl_names[[x]]),
     tbl_names = tbl_names,
     obj = d
   ) %>% rlang::set_names(keep_sets)
@@ -123,7 +125,7 @@ create_para_input <- function(paras, n_samples, d) {
     result <- expand.grid(paras) %>% tibble::as_tibble()
     result <- result %>%
       dplyr::left_join(
-        suppressWarnings(d %>% tidyr::nest(-.data$sample_id, .key = "data")),
+        d %>% tidyr::nest(data = -.data$sample_id),
         by = "sample_id"
       )
   }
@@ -188,12 +190,12 @@ combi_run_curvep_in <- function(d, mask, n_samples, keep_sets, paras) {
 #' @noRd
 #'
 flat_result_tbl <- function(d, keep_set) {
-  suppressWarnings(result <- d %>%
+  result <- d %>%
     dplyr::mutate(
       temp = purrr::map(.data$rcurvep_obj, ~ .x[['result']][[keep_set]])
     ) %>%
     dplyr::select(-.data$rcurvep_obj) %>%
-    tidyr::unnest())
+    tidyr::unnest("temp")
 
   return(result)